refactor(cli): flatten CLI package layout
