+++ go/core/cli/internal/commands/agentinstance/lifecycle_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/google/uuid"
 	apiv1alpha1 "github.com/kagent-dev/kagent/go/api/gen/kagent/api/v1alpha1"
-	clioutput "github.com/kagent-dev/kagent/go/core/cli/internal/cli/output"
+	clioutput "github.com/kagent-dev/kagent/go/core/cli/internal/output"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
