+++ go/core/cli/internal/commands/bug_report.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"fmt"
