+++ go/core/cli/internal/commands/const.go
@@ -1,4 +1,4 @@
-package cli
+package commands
 
 import (
 	"os"
