+++ go/core/cli/internal/commands/dashboard.go
@@ -1,6 +1,6 @@
 //go:build !darwin
 
-package cli
+package commands
 
 import (
 	"context"
