+++ go/core/cli/internal/commands/dashboard_darwin.go
@@ -1,6 +1,6 @@
 //go:build darwin
 
-package cli
+package commands
 
 import (
 	"context"
