+++ go/core/cli/internal/commands/mcp/add_tool.go
