+++ go/core/cli/internal/commands/mcp/build.go
