+++ go/core/cli/internal/commands/mcp/build_test.go
