+++ go/core/cli/internal/commands/mcp/deploy.go
