+++ go/core/cli/internal/commands/mcp/init.go
