+++ go/core/cli/internal/commands/mcp/init_test.go
