+++ go/core/cli/internal/commands/mcp/inspector.go
