+++ go/core/cli/internal/commands/mcp/integration_test.go
