+++ go/core/cli/internal/commands/mcp/root.go
