+++ go/core/cli/internal/commands/mcp/run.go
