+++ go/core/cli/internal/commands/mcp/run_test.go
