+++ go/core/cli/internal/commands/mcp/secrets.go
