+++ go/core/cli/internal/output/output.go
