+++ go/core/cli/internal/output/output_test.go
